More organisation, more code mocking-up
The goal is to try to envision how data will be laid out from end-to-end, and how the important parts of compatibility targets will map to Impure's systems. GZDoom takes priority by virtue of being massively more prolific, but room can always be made for Eternity, not to mention that much of this is liable to become consigned to scripts.

Also, organise all main-loop-critical structures into one struct.

Next on the docket is to construct levels from WAD contents.