Start on networking backend for dedicated server
This is a good place to explore how these systems are going to look, since the surrounding infrastructure is nothing more than a CLI. This can later be abstracted into engine structures and refactored for use in the client too.

The fasthash crate is imported here but ended up going unused for now. It'll later be helpful in sync checks for lockstep multiplayer, so I'll leave it in the manifest.