Better Lua logging
The 4 global Lua log functions now take a variable number of arguments that can be formatted using "{}" like in Rust, powered by the formatx library. Support for more advanced formatting may come later. This is with the help of a universally-exported global function `repr` - a shortcut for Serpent-based pretty-printing - which, in my experience with SCFA, will prove extremely useful with time.