Beginnings of ZScript transpiler
Start using Gutawer's parser to generate ASTs, and report on parse errors in a way that makes it easy for the user to find and debug them. The parser itself still trips up on some code which is valid to GZDoom's parser, so that's going to need fixing before transpilation work can continue.