Key `DataCore::asset_map` via `u64` instead of a string
Since the rest of both native- and script-side code will pass asset IDs to specialised functions, we can always rely on having an abstraction layer that's generic over `A: Asset`. As such, we can reliably mix the asset's domain string into an incoming ID, so modders don't have to concern themselves with it.