Make `DataCore::add` fallible, fix its lump-mapping behaviour
For reasons I can't explain, this was inserting stringified asset indices as lump map keys. Now lump names are properly extracted - split on the first period and only the first 8 characters. Return an error (a variant added to a now-generalised `AssetError`) if an attempt is made to clobber an existing ID under the same namespace, or to add to a namespace that doesn't exist.