Start building Impure's implementation of Action Code Script
This is going to be a painful and tedious process (but absolutely necessary) that might have ramifications on the remainder of the project. It also has the benefit of not demanding any transpilation, meaning that even if DECORATE/ZScript compatibility ultimately fail, this has a better chance of still working. I'll try to keep commits small going forward, so this is little more than a module with some skeletal structures and enumerations.