Remove Teal
After some consideration, Teal isn't the way forward for Impure. Its type system is not very powerful and its language support is not up to the same level of quality as Lua. Adding the extra complexity of a compilation step is not worth the upsides when the same advantages can be gained through static analysis from good tooling.