Put contents of /data/ into /data/impure, merge with Lua libdefs
This creates a unified directory that can be stripped and packaged for inclusion in any user's workspace for the language server, containing both documents and implementations for reading at the user's leisure. It also has the added benefits of opening the data directory up to holding the contents for other official packages than just the engine's data and enables the language server to retrieve definitions given just a path alone, improving the experience for any developer working on the engine's scripts.