LithScript language: parsing, part 1
My mind is made up. Lua has to go, not just because it lacks features like a compile-time scope system and that it would be too foreign to users migrating from GZDoom, but also because of the inherent performance cost of a dynamically-typed language at the native interop boundary.

Having evaluated all the other options, I've decided I'm just going to try building an entirely new successor to ZScript on top of cranelift. This is, I think, the best option for the following reasons:
- Ease of transpilation from existing ZDoom code.
- Leverage some prior art, like from ZScript itself, or Gutawer's ZScript parser, or zsdoc.
- Maximum fulfillment of bespoke needs, minimum baggage from other languages, maximum ability to combine the best features and implementations from other languages.
- At a moment's notice I can strip out the ZScript-esque parts of the language and bring it to another project.
- https://justforfunnoreally.dev/

The first step is a parser, generated via Pest. The grammar will be built up over following commits piece-by-piece; for now it's just some foundationals. There's also an array of reserved keywords.