Decouple `FileSpan` from any input, rename it to `Span`
The reference within not only made the borrow checker very upset when trying to implement the parser, but also inextricably tied the whole Lith language to pest.

Having the only `Span` type (the ZS parser along with its span type is soon going to another crate) be tied to a VFS entry is also undesirable, since there will certainly be a need to parse strings that don't come from the VFS at all.