Foundations for custom ECS
Shipyard looks like a good library altogether but it's not suited to this project. Even with `unsafe`, a fast boundary between any given scripting language and the ECS is, as best as I can tell, impossible. With an in-house solution, the indirections needed to make an ECS dynamic over any possible user-defined types is obviated and one can access an entity's component nearly as quickly as dereferencing a pointer.

Shipyard will remain a dependency since I'm going to experiment with using it to implement GUIs.

- `mlua` `UserData` implementations of existing components have been removed since Lua is soon going to be removed from Impure.
- Existing components have been moved to a new module to keep the top-level module clean(er).