Rename from "Impure" to "VileTech"
It was never a very good name to begin with. With an out-of-the-box configuration, I would prefer for this engine to be even more suited to purists than GZDoom is, what with its myriad modifications to vanilla. Besides, "impure" is probably a good name for someone else's Rust crate that allows an interior mutability pattern or whatnot. Now, we have a name that's much more memorable, distinctly Doom, and better for SEO.