Overhaul developer GUI architecture in the client
One window anchored to the screen top, split into two horizontally-resizable panels. Each panel can be swapped for a different menu. Resembles GZDoom's console, but with some of the customizability and ease of Blender's GUI system.

A basic dev-GUI for the audio core has been written, with adjustments made to the core itself to acommodate.