Bring back SoundFont collection code
ZMusic is eventually going to go. It isn't a good fit for this project's needs; for that matter, it's not a good library at all. Eventually there will be a custom implementation with nodi, midly, and fluidlite. For now, get this code down so the user will eventually be able to add and choose their own SoundFonts.