Replace ZMusic with nodi/midly/fluidlite-based solution
Still kira-integrated. Much of the functionality we previously had will have to be written -- a MUS parser is mandatory, for instance -- but I would have to do that anyway in C++. In the (likely) event that this project never comes to fruition, any MIDI tech I wrote for it can be re-used by someone else.