Prevent client lock-up if core fails to construct
The console message receiver drops and the sender, to the best that I can tell, causes a deadlock when it tries to use the poisoned channel. Another possible solution to use a ref-counted pointer to ensure the receiver lives long enough to oversee reporting on the core init error, but I won't do this until I see a necessity for it.