Start migrating to Bevy
Here's a big one. I had written the possibility of doing this off for a while, given the incompatibility of Bevy's ECS with VileTech's particular needs, but I've recently made two observations:
- Using Bevy will give me access to orders of magnitude more technology than not doing so.
- I don't have to use Bevy's ECS for the gameplay code; I can just use it as a renderer.

As such, I'm just going to go through with this. The first step is just re-arranging Cargo manifests; I'd like to try to keep this next few commits somewhat small. The next thing to do is get back to parity with the non-Bevy client and then start trying to draw map geometry.

Having removed `renet` in favour of a `quinnet` Bevy add-on, the dedicated server code got pared down.