Use `time` instead of `chrono` for `engine/build.rs`
See https://github.com/jerome-trc/viletech/security/dependabot/1. For whatever reason nobody with push access to `chrono` has resolved this upstream yet. This reduces the dependency size of the project, and upon fresh review it had no real reason to be a part of the engine, so `time` is only a build dependency.