Implement game load cancellation
The catalog is not in an entirely coherent state now (e.g. its ability to revert itself upon failure has been torn out), but only because, again, the catalog needs to change. Either file reading needs to be lazier, so that less time is wasted reading binary files that might not get used before script compilation is done, or there needs to be memoization between loads.