Remove all uses of the `#[inline]` attribute
No reason to start applying this until the time comes to start profiling and micro-optimizing the hottest of the hot code.

Its use in the SIMD code in particular was based on the usage of `__forceinline` by Gaijin Entertainment in Dagor's vector math library, and while I presume that they tested it to be better, there's no reason to assume off-hand that rustc will exhibit the same performance characteristics as their C++ toolchains.