Redo how licensing and attribution are documented
Mostly to ensure that proper notices are attached to the repository in a way that can, if need be, be distributed along with the final binary. This also reduces duplication in the source.

It seems redundant to list off boilerplate for every crate I link against when there are no licensing requirements to do so and any curious party can easily track that information down on their own, so that was removed.

Additionally, add a couple of extra names to the "Special Thanks" section.