Building up sim's level code
I was previously not planning on leveraging Bevy's ECS in this way - I don't want to fall into the trap of trying to use ECS as a solution to every single problem - but I feel obligated to at least experiment with it to know for certain whether it's suitable. If nothing else, I except Bevy's ECS to be better than anything I could implement on my own.

This work is at a "sketching" stage. The immediate next step is to determine how level geometry will be converted into render-ready meshes, and writing custom shaders.