VZS: start overhauling language, toolchain
The plan going forward is to build the language in phases; implement a layer of semantics (functions, annotations, and basic arithmetic first), experiment to make it function correctly and ergonomically at both the syntactic and behavioural levels, and then write it into the spec, and then repeat with a new phase (e.g. structures, unions).

- Catch up to doomfront's breaking changes.
- Add `cranelift-interpreter`; bring Cranelift versions to 0.95.1. When I started work on the original VZS node-based interpreter, I wasn't aware that an interpreter for CLIF existed. I'd like to see what it can do; I don't expect it to be highly performance-oriented, since it seems to be foremostly for testing Cranelift's behaviour, but it will benefit from all of Cranelift's optimizations without any extra work on my part, so it may be worthwhile when JIT is unavailable or unnecessary.