Vendor DoomFront
This crate was previously kept in another of my repositories, doom-rs. What I've recently realised, however, is that it had a problem; my upcoming DoomLSP would depend on it, like VileTech, but both would rely on toolchain layers that I had previously marked as being out-of-scope for DoomFront itself.

Thus, DoomFront is becoming a part of VileTech (which I want to rebrand as a family of Doom technologies) and getting its scope expanded to include semantic checking and IR generation, making it truly a set of frontends.

Since doom-rs is getting closed - and Stardate and Subterra always had questionable reason to exist - references to those two are gone too.

This kind of work likely needs to go further, since having libraries or applications depend on one monolithic crate that also brings in all of Bevy isn't practical.