Begin migrating VZScript into its own crate
Better for applications that don't want to link against the entire engine, especially if those applications prove to not even be Doom-related.

This is an extremely bare-bones commit - not even the few parts of the old frontend were brought over - since DoomFront is getting some breaking changes soon and VZScript's file system is getting a makeover to a more orthodox file-per-module and library framework.

Note that Cranelift has been updated to 0.96.1.