Start overhauling VFS in its own new crate
Since this depends on types that don't belong in it, `utils` has also gotten its own crate.

The big change is that files will be read lazily rather than eagerly, so that no more memory is consumed than necessary. It will also be possible to cache a file, allowing the best of both worlds.

Additionally:
- Files are stored in a `HashMap` rather than an `IndexMap` to eliminate an indirection in the general case.
- Directory `IndexSet`s are boxed to not bloat the size of all other files.