Finish migration of VFS code to new crate
In the end, I opted to not going through with a lazy-read system. Not only would doing so have increased code complexity up front, but it also would have required non-trivial space to support archive nesting the way that the current code does. It's not even necessary for feature parity with GZDoom, since it also "locks in" the state of files within WADs within zips, for example.

The only missing pieces now are re-reads and duplication of memory between large files and data objects, but I have solutions in mind for both.