DoomFront: remove `zdoom::Version`
As it turns out, I badly misunderstood how Logos callbacks work. I thought a skipped match would apply rollback, but this was not happening, and it was breaking parses expecting identifiers which overlapped with keywords. Parsers will just have to deal with versioned keywords on their own, but I'll get this done over future commits.