DoomFront: experimental hand-written LANGUAGE parser
I had not intended to try hand-writing resilient recursive descent parsers for every DSL within DoomFront's scope, but then Alexey Kladov released a tutorial for doing exactly that and made it look easy.

The advantages of this approach are:
- Massively reduced compile times. No more compiler RAM blow-ups.
- Parsers like these are much more amenable to step-through debugging.
- No more incomprehensible error messages.
- No cycles spent on parser construction and re-construction.
- Simpler "combinator" function signatures; no need to generate elaborate return types with macros.
- Smaller dependency tree and smaller binaries.

None of that is to mention the performance; I observed a 7.65 times speed increase with one cursory benchmark. These aren't even extremely hard to write, just slightly more tedious and verbose. The tricky part will be left-recursion, but I have resources to read up on for getting that to work. In following commits, I'll try to make this more concise to use and add error recovery and green cache support, and then re-assess. Maybe this becomes DoomFront's final future direction after all.