Lith: start migrating old parser code to rust-peg
A hand-written parser is more resistant to changes in the grammar, since the entire state machine is explicit, making them less suited to working on languages whose syntax is in flux. rust-peg, meanwhile, may not have any resilience, but this isn't very important at this stage in Lith's life. Every time a LithScript "phase" gets crystallized, a hand-written parser can be written for it, while the rust-peg version moves on to further experiments.