Lith: revise compilation model
I somehow got it in my head that it made sense for a source file to be compiled to a Lith module, but upon testing, this absolutely cannot work due to linking needs (e.g. it becomes impossible for two functions to call each other). Instead, it makes more sense to do things the Rust way and have the library be the atomic compilation unit.

- What was a Lith "module" is now a "container" to distinguish it from Cranelift's modules.
- Native symbols are registered on a per-library basis.
- The old runtime abstractions over libraries and modules is gone until I know what codegen will output (it will definitely need to emit RTTI that a caller can look up).
- New src/compile.rs module to clean up and generalise src/viletech.rs.