Lith: still more sketching about compilation, transpilation
I'm still not yet certain that this is a satisfactory model, but it's much closer. Mind that the state of the code as of this commit is a disaster area, since I'm trying to get a broad look at every part of this at once so I can tease out flaws in the architecture before specifying any piece too deeply.

- LIR now forms a flat structure, with indices used for indirection instead of boxes.
- Redo runtime information abstractions (i.e. the `Symbol` trait and everything surrounding it), with a focus on type definition/RTTI, to better suit the compiler's needs. This includes the new `project::Library`, emitted at the end of compilation.
- Some basic issue report generation.
- Constrain `Send and `Sync` impls for `TPtr`.
- Use the `smallvec` macro less.
- What was `lir::Name` is now `compile::QName`.