Begin migrating from LithScript to VZScript
Back and forth and back and forth.

Working on Lith has left me with the conclusion that it is a lose-lose proposition. The language and its toolchain become tortured to accommodate the needs of GZDoom compatibility, while the users have a harder hypothetical path of transitioning from the old languages to the new due to Lith's divergence.

VZScript is back and here to stay. It will still be syntactically incompatible but, in a novel turn, it will be possible to include VZS files in ZS files and vice-versa.