VZS: move back to Cranelift from MIR
This was an interested experiment, but what I realized was that Cranelift's existing interpreter capabilities are more suited to my needs (or at least will be with some gentle modification) for compilation and debug step-through. Support for compiling C is not important (honestly, there will probably be a Cranelift-based compiler available eventually) and I can turn VIR into C and then send it to Clang and GCC to verify it thus.