Start moving to a Z-up coordinate system
Whatever wrangling with Bevy's Y-up coordinates is necessary, it will be worthwhile to increase cohesion across the rest of the code, especially with foreign code that gets ported in.

VileData in particular ought to work the same way as the original doom.exe since it's a public interface.

This commit is deliberately not thorough, since the "sim" code is getting re-done soon.