Re-orient viletech-acs
For the foreseeable future, this will just be an object file reader and corresponding structures. If there ever comes a time it's able to take in ACS source, that's years away.

Meanwhile, VZS is going to take on the role of JIT-compiling from bytecode. The compile pipeline is being adjusted for this accordingly.