Lith: moving towards better CTFE
The crux of this change is expanding the information available to compile-time functions to support future compile built-ins, such as one which can retrieve the default value for a parameter if possible. The signature for a function like this has been unified with the signature that an embedder would use to register their own functions; accodingly, some first steps have been taken in exposing more compiler internals. This will eventually be taken further to allow for better third-party tools, more inspection capabilities by the embedder at different parts of the compile pipeline, and for selective generation of run-time information, so that only what the embedder needs gets allocated.