Update Bevy from 0.11.2 to 0.12.0
Additionally necessary updates include:
- `egui`, 0.22.0 to 0.23.0
- `bevy_egui`, 0.21.0 to 0.23.0
- `bevy_quinnet`, 0.5.0. to 0.6.0

Lithica's design should be at least somewhat informed by the design of the engine's game sim, and the design of that is completely informed by whether Bevy is suitable for our purposes. Now that Asset V2 is out, it's time to really put it to the test.