VFS: sketching out yet another overhaul
Back to basics in a big way. The new VFS is not as eager about ingesting file content; instead, the necessary structures are put in place for reading all those bytes later, to reduce the amount of time necessary to start the game. Then, more content can be loaded in the background (e.g. prepare MAP02 will user is playing MAP01).