Lith: link against cwal-rs to prepare for next steps
The language is going to be pivoting significantly to take after Terra rather than Zig. This is mostly because of how beneficial it would be to have a dynamic component with which users can write quick scripts (e.g. use the CLI to mount files to a VFS and then perform operations on them), but it would also open the door to easier transpilation from ZDoom languages and be better for the loading phase of the VTGE.