Engine: don't send less than 2 points to `points_to_poly`
This is a mildly rare case that still needs to be addressed, maybe heuristically. There's no strict need to `debug_assert` in the function itself, because the behaviour is coherent even if only two points are passed, but the output will never pass the following filter, so it's just wasted cycles.