Initial commit for new branch: `zmsx`
I'd like to revisit this matter to see if, with some not-too-invasive changes, ZMusic can become a suitable MIDI playback solution for VileTech in the long term.

To start with, this revision is a straightforward vendor of the ZMusic source at https://github.com/ZDoom/ZMusic/commit/5ce525efc75a97d1391b46cca4def41a13ff1619.