Initial commit for new branch: `zdfs`
Investigating whether this can be wrapped up for VileTech's use. It doesn't look like it will be able to cover both VTGE and VTEd out-of-the-box but perhaps it can be modified for those. If not, it has the potential to be useful to somebody else, and understanding it better will be valuable.

There's no meaningful C API to use (or bind to) yet; it's just a straightforward vendor of the source that barely builds.