Disable the source port-related parts of the workspace for now
I don't think that retracing all the steps GZDoom made to get to its current point is sensible; the hypothetical end product of this endeavour would not even be extremely useful to me, let alone others. Work will instead proceed on FFI crates, DoomFront, and DoomLS, since the value of these efforts is apparent.

Note that this includes Lithica, on which work will proceed in a repository of its own.