Purge dsda-doom code
Now that I have a private dsda-doom fork which satisfies my Doom daily-driving needs for the foreseeable future, I'm not in any hurry to get VileDoom to a usable state. As such, I'm instead going to slowly build up both VileDoom and Hellblood from scratch, and just enjoy the process.